Move parsing stuff under tabular import.
