+++ spatialprofilingtoolbox/workflow/tabular_import/parsing/cell_manifest_set.py
