+++ spatialprofilingtoolbox/workflow/tabular_import/parsing/cell_manifests.py
