+++ spatialprofilingtoolbox/workflow/tabular_import/parsing/diagnosis.py
