+++ spatialprofilingtoolbox/workflow/tabular_import/parsing/study.py
